requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
Werkzeug==2.2.2
numpy~=1.23.5
julia~=0.6.0